Move a few vendor files over to the lib directory
This is what got moved:

- [x] jello.js
- [x] memory_store.js
- [x] number_to_human.js
- [x] session.js
- [x] time_ago_in_words.js

So these files are more lib than vendor and probably could be pushed out
to their own repositories... something to do in the future. The other
added benefit is GitHub's file search ignores the `/vendor` directory so
we never see these files in their enviornment. This will fix that issue.

The files that remain in `/vendor` are fine remaining for now. We should
eventually get Justin to setup embetter as a npm package and create a PR
for Honeybadger to support server rendering.
