+++ src/components/editor/InputComponent.js
@@ -1,7 +1,7 @@
 import { addKeyObject, removeKeyObject } from '../viewport/KeyComponent'
 import { getWordFromSelection } from './SelectionUtil'
 import { emojiRegex, userRegex } from '../completers/Completer'
-import { isFirefox } from '../../vendor/jello'
+import { isFirefox } from '../../lib/jello'
 
 const methods = {}
