+++ src/containers/CommentToolsContainer.js
@@ -11,7 +11,7 @@ import { openModal, closeModal } from '../actions/modals'
 import ConfirmDialog from '../components/dialogs/ConfirmDialog'
 import FlagDialog from '../components/dialogs/FlagDialog'
 import { getEditorId } from '../components/editor/Editor'
-import { scrollToLastTextBlock } from '../vendor/jello'
+import { scrollToLastTextBlock } from '../lib/jello'
 import { CommentTools } from '../components/comments/CommentTools'
