+++ src/containers/EnterContainer.js
@@ -5,7 +5,7 @@ import { replace } from 'react-router-redux'
 import debounce from 'lodash/debounce'
 import sample from 'lodash/sample'
 import set from 'lodash/set'
-import { isAndroid, isElloAndroid } from '../vendor/jello'
+import { isAndroid, isElloAndroid } from '../lib/jello'
 import { ONBOARDING_VERSION } from '../constants/application_types'
 import { FORM_CONTROL_STATUS as STATUS } from '../constants/status_types'
 import { selectCurrentStream, selectCoverDPI } from '../selectors/gui'
