+++ src/containers/ForgotPasswordContainer.js
@@ -2,7 +2,7 @@ import React, { Component, PropTypes } from 'react'
 import { connect } from 'react-redux'
 import shallowCompare from 'react-addons-shallow-compare'
 import sample from 'lodash/sample'
-import { isAndroid } from '../vendor/jello'
+import { isAndroid } from '../lib/jello'
 import { FORM_CONTROL_STATUS as STATUS } from '../constants/status_types'
 import { selectCoverDPI } from '../selectors/gui'
 import { selectPromotionsAuthentication } from '../selectors/promotions'
