+++ src/lib/memory_store.js
