+++ src/lib/number_to_human.js
